fix: 상품 상세 조회 응답값 수정 1. 필드 이동 2. Sku 도메인 이동
