+++ src/main/java/com/tutti/server/core/product/payload/response/ProductOptionResponse.java
@@ -10,7 +10,11 @@ public record ProductOptionResponse(
         String firstOptionValue,
         String secondOptionName,
         String secondOptionValue,
+        int sellingPrice,
+        int discountPrice,
         int additionalPrice
+//        int stockQuantity,
+//        boolean almostOutOfStock
 ) {
 
     public static ProductOptionResponse from(ProductItem productItem) {
@@ -20,7 +24,11 @@ public record ProductOptionResponse(
                 .firstOptionValue(productItem.getFirstOptionValue())
                 .secondOptionName(productItem.getSecondOptionName())
                 .secondOptionValue(productItem.getSecondOptionValue())
+                .sellingPrice(productItem.getSellingPrice())
+                .discountPrice(productItem.getDiscountPrice())
                 .additionalPrice(productItem.getAdditionalPrice())
+//                .stockQuantity(sku.getStockQuantity())
+//                .almostOutOfStock(sku.getStockQuantity() <= 5)
                 .build();
     }
 }
