+++ src/main/java/com/tutti/server/core/sku/domain/SkuHistoryType.java
@@ -1,0 +1,8 @@
+package com.tutti.server.core.sku.domain;
+
+public enum SkuHistoryType {
+
+    IN,         // 입고 - 입고에도 대량 입고, 반품으로 인한 입고는 구분하지 않음
+    OUT         // 출고
+
+}
