--- src/main/java/com/tutti/server/core/stock/domain/SkuHistoryType.java
@@ -1,8 +1,0 @@
-package com.tutti.server.core.stock.domain;
-
-public enum SkuHistoryType {
-
-  IN,         // 입고 - 입고에도 대량 입고, 반품으로 인한 입고는 구분하지 않음
-  OUT         // 출고
-
-}
